initial prototype
initial prototype for serving tiles through
separate isolates. This allows for concurrent
rendering of tile images, without using the UI
thread. The intention is to reduce jank caused
by work on the UI thread or on the raster thread.
By performing rendering in isolates launched
with flutter_isolate, we are able to render tiles
separately from the main process.

This prototype is pretty rough, with notable
defficiencies:
- isolates are leaked whenever the tile provider
  is recreated
- data transfer between isolates causes each tile
  to incur about 2s of latency, which in turn
  causes queueing
- queueing causes additive latency
- conversion of raster images to png occurs twice